2nd relay now is always on GND by default
This protects the load in case there's a malfunction in the supply - the
load will have no power by default.
There's also a pull-up resistor on the SC protection because of that
(explanation is on the schematic).

Also changed the noise voltage source to 500 Hz instead of 5 kHz,
because it's more realistic.